// mebel-next/libs/enums/product.enum.ts
export enum ProductType {
	SOFA = 'SOFA',
	CORNER_SOFA = 'CORNER_SOFA',
	ARMCHAIR = 'ARMCHAIR',
	BED = 'BED',
	POUF = 'POUF',
	MATTRESS = 'MATTRESS',
	KIDS = 'KIDS',
}

export enum ProductStatus {
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}

export enum ProductMaterial {
	FABRIC = 'FABRIC',
	VELVET = 'VELVET',
	MICROFIBER = 'MICROFIBER',
	LEATHER = 'LEATHER',
	ECO_LEATHER = 'ECO_LEATHER',
}

export enum ProductLocation {
	TASHKENT = 'TASHKENT',
	SAMARKAND = 'SAMARKAND',
	BUKHARA = 'BUKHARA',
	ANDIJAN = 'ANDIJAN',
	FERGANA = 'FERGANA',
	NAMANGAN = 'NAMANGAN',
	NAVOI = 'NAVOI',
	KASHKADARYA = 'KASHKADARYA',
	SURKHANDARYA = 'SURKHANDARYA',
	KHOREZM = 'KHOREZM',
	JIZZAKH = 'JIZZAKH',
	SIRDARYA = 'SIRDARYA',
	KARAKALPAKSTAN = 'KARAKALPAKSTAN',
}